requirements.txt
pdjango
django-cors
django-cors-headers